refactor(v1-release): extract Dialog autofocus into useAutofocusOnOpen
Move the open-time `[autofocus]` handling out of Dialog.vue into a
reusable composable, and scope the container lookup to the dialog's own
content element so nested dialogs focus their own marker rather than
the first one in document order. Also drops a stray debug log.
